Agora deve dar ig

+++ laravel/app/Http/Controllers/api/AuthController.php
@@ -4,11 +4,11 @@ namespace App\Http\Controllers\api;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use App\Http\Requests\LoginRequest;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
-use App\Http\Requests\Auth\LoginRequest;
 use App\Models\User;
-use Illuminate\Support\Str; // Add this at the top for the Str helper
+
 
 class AuthController extends Controller
 {
@@ -16,16 +16,14 @@ class AuthController extends Controller
     {
         // Only deletes if token expired 2 hours ago
         $dateTimetoPurge = now()->subHours(2);
-        DB::table('personal_access_tokens')
-            ->where('expires_at', '<', $dateTimetoPurge)->delete();
+        DB::table('personal_access_tokens')->where('expires_at', '<', $dateTimetoPurge)->delete();
     }
+
     private function revokeCurrentToken(User $user)
     {
         $currentTokenId = $user->currentAccessToken()->id;
         $user->tokens()->where('id', $currentTokenId)->delete();
     }
-
-
 
     public function login(LoginRequest $request)
     {
@@ -34,24 +32,23 @@ class AuthController extends Controller
         if (!Auth::attempt($credentials)) {
             return response()->json(['message' => 'Unauthorized'], 401);
         }
-        $token = $request->user()->createToken('authToken', ['*'],
-            now()->addHours(2))->plainTextToken;
+        $token = $request->user()->createToken('authToken', ['*'], now()->addHours(2))->plainTextToken;
         return response()->json(['token' => $token]);
     }
+
     public function logout(Request $request)
     {
         $this->purgeExpiredTokens();
         $this->revokeCurrentToken($request->user());
         return response()->json(null, 204);
     }
+
     public function refreshToken(Request $request)
     {
         // Revokes current token and creates a new token
         $this->purgeExpiredTokens();
         $this->revokeCurrentToken($request->user());
-        $token = $request->user()->createToken('authToken', ['*'],
-            now()->addHours(2))->plainTextToken;
+        $token = $request->user()->createToken('authToken', ['*'], now()->addHours(2))->plainTextToken;
         return response()->json(['token' => $token]);
     }
-
 }
